コンフリクト解消、画面遷移、クラッシュ問題解消
