+++ app/src/main/AndroidManifest.xml
@@ -44,5 +44,8 @@
         <activity
             android:name=".RegisterActivity"
             android:exported="true" />
+        <activity
+            android:name=".MorningRoutineActivity"
+            android:exported="true" />
     </application>
 </manifest>
